refactor: consolidate checking for role functions
