+++ app/pages/sign-in/index.tsx
@@ -6,7 +6,7 @@ import { Loading } from "src/components/general/Loading";
 import { SignIn } from "src/components/sign-in";
 import { DeveloperModeContext } from "src/state/developer-mode/context";
 import { SessionContext } from "src/state/session/context";
-import { checkIsAuthenticated } from "src/utils/checkIsAuthenticated";
+import { checkIsAuthenticated } from "src/utils/checkIsRoleExist";
 
 const SignInPage = () => {
   // context
