+++ app/pages/volunteers/index.tsx
@@ -5,8 +5,10 @@ import { SignIn } from "src/components/sign-in";
 import { Volunteers } from "src/components/volunteers";
 import { DeveloperModeContext } from "src/state/developer-mode/context";
 import { SessionContext } from "src/state/session/context";
-import { checkIsAuthenticated } from "src/utils/checkIsAuthenticated";
-import { checkIsCoreCrew } from "src/utils/checkIsCoreCrew";
+import {
+  checkIsAuthenticated,
+  checkIsCoreCrew,
+} from "src/utils/checkIsRoleExist";
 
 const VolunteersPage = () => {
   // context
