--- app/src/utils/checkIsAuthenticated.ts
@@ -1,13 +1,0 @@
-import { ACCOUNT_TYPE_ADMIN, ACCOUNT_TYPE_AUTHENTICATED } from "src/constants";
-import { IAccountTypePayload } from "src/state/developer-mode/reducer";
-
-export const checkIsAuthenticated = (
-  { isEnabled, value }: IAccountTypePayload,
-  isAuthenticatedSession: boolean
-) => {
-  return (
-    (isEnabled &&
-      (value === ACCOUNT_TYPE_ADMIN || value === ACCOUNT_TYPE_AUTHENTICATED)) ||
-    (!isEnabled && isAuthenticatedSession)
-  );
-};
